% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adjustment.r
\name{createthemodel}
\alias{createthemodel}
\title{create the model using caret package}
\usage{
createthemodel(data, train_index, method)
}
\arguments{
\item{data}{which you want to build the model for}

\item{train_index}{theis index was created using machine_learning function and it is required to make sure
that we train and test the data on the same set}

\item{method}{which method you want to train the model such as 'lda' and 'gbm'}
}
\description{
This function train the model using 80% of the data and retrive the ROC using the 20% of the data
}
\examples{
createthemodel(data,train_index,method)
createthemodel(data,train_index,method)
}
\keyword{creat}
\keyword{model}
\keyword{the}
