% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data.hmdb}
\alias{data.hmdb}
\title{data.hmdb}
\format{data.frame
$ Accession   : Factor w/ 53287 levels 'HMDB0000001',..: 1 1 2 2 2 2 3 3 3 3 ...
$ Name        : Factor w/ 53287 levels '(±)-Glycerol 1,2-diacetate',..: 45 45 55 55 55 55 146 146 146 146 ...
$ Pathway_Name: Factor w/ 48607 levels '11-beta-hydroxylase deficiency (CYP11B1)',..: 46317 46318 98 23433 46265 48580 19 23491 46174 46176 ...}
\source{
\url{http://www.hmdb.ca/system/downloads/current/hmdb_metabolites.zip}
}
\usage{
data.hmdb
}
\description{
data.hmdb
}
\keyword{datasets}
