% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data.metaboAnalyst}
\alias{data.metaboAnalyst}
\title{data.metaboAnalyst}
\format{data.frame
Classes ‘tbl_df’, ‘tbl’ and 'data.frame': 183937 obs. of  8 variables:
$ hmdb_id   : chr  'HMDB0000001' 'HMDB0000001' 'HMDB0000001' 'HMDB0000001' ...
$ name      : chr  '1-Methylhistidine' '1-Methylhistidine' '1-Methylhistidine' '1-Methylhistidine' ...
$ synonym   : chr  '(2S)-2-amino-3-(1-Methyl-1H-imidazol-4-yl)propanoic acid' ...
$ pubchem_id: int  92105 92105 92105 92105 92105 92105 92105 92105 92105 92105 ...
$ chebi_id  : int  50599 50599 50599 50599 50599 50599 50599 50599 50599 50599 ...
$ kegg_id   : chr  'C01152' 'C01152' 'C01152' 'C01152' ...
$ metlin_id : int  3741 3741 3741 3741 3741 3741 3741 3741 3741 3741 ...}
\source{
\url{http://www.metaboanalyst.ca/MetaboAnalyst/faces/docs/Resources.xhtml}
}
\usage{
data.metaboAnalyst
}
\description{
data.metaboAnalyst
}
\keyword{datasets}
