% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data.smpdb}
\alias{data.smpdb}
\title{data.smpdb}
\format{data.frame
'data.frame': 19245 obs. of  2 variables:
$ pathway: Factor w/ 705 levels '11-beta-hydroxylase deficiency (CYP11B1)',..: 39 39 39 39 39 39 39 39 39 39 ...
$ hmdb_id: chr  'HMDB0000538' 'HMDB0000161' 'HMDB0000045' 'HMDB0000250' ...}
\source{
\url{http://smpdb.ca/downloads/smpdb_metabolites.csv.zip}
}
\usage{
data.smpdb
}
\description{
data.smpdb
}
\keyword{datasets}
