% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresSelection.r
\name{featuresSelection}
\alias{featuresSelection}
\title{A featuresSelection Function}
\usage{
featuresSelection(PDSmatrix, threshold = 0.5, method = "info")
}
\arguments{
\item{PDSmatrix}{from PDSfun function}
}
\description{
This function allows you to reduce the pathway diemsion using xxxx
}
\examples{
selected_Pathways_Weka= featuresSelection(PDSmatrix)
featuresSelection(PDSmatrix)

}
\keyword{features}
\keyword{selection}
