% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{GetMappingResultTable}
\alias{GetMappingResultTable}
\title{A GetMappingResultTable Function}
\usage{
GetMappingResultTable(metaboliteDataSet)
}
\description{
This function allows you to generate csv file with user metabolites matched names, kegg, pubchem
and hmdb ids
This function was modified version of the name.match function in the below link:
https://github.com/cangfengzhe/Metabo/blob/master/MetaboAnalyst/website/name_match.R
}
\examples{
GetMappingResultTable()
GetMappingResultTable()

}
\keyword{data.map}
