% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetabolitePathways.R
\name{getMetabolitePathways}
\alias{getMetabolitePathways}
\title{Takes a list of metabolite names or IDs and adds pathways from the HMDB database.}
\usage{
getMetabolitePathways(metaboliteDataSet)
}
\description{
Takes a list of metabolite names or IDs and adds pathways from the HMDB database.
}
