% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{IDMapping}
\alias{IDMapping}
\title{A IDMapping Function}
\usage{
IDMapping(sourceDatabase, metaboliteDataSet)
}
\arguments{
\item{sourceDatabase}{which is one of kegg id, pubchem is, hmdb id}
}
\description{
This function allows you to exact matching of your metabolites hmdb, pubchem and kegg ids with data base
consists of more than 20k metabolites.
make sure that you have the below database for the exact matching:
cmpd_db.rda
This function was modified version of the name.match function in the below link:
https://github.com/cangfengzhe/Metabo/blob/master/MetaboAnalyst/website/name_match.R
}
\examples{
IDMapping(sourceDatabase)
IDMapping()

}
\keyword{Exact}
\keyword{Matching}
