% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.adjust_model.r
\name{lilikoi.adjust_model}
\alias{lilikoi.adjust_model}
\title{Model Adjustment method using clinical factors}
\usage{
lilikoi.adjust_model(mlResults, PDSmatrix, selected_Pathways_Weka,
  metaboliteMeasurements, clinical_factors_data, factors)
}
\arguments{
\item{mlResults}{The models and other results generated from machine_learning method}

\item{PDSmatrix}{the PDS matrix generated using lilikoi.get_pd_scores function}

\item{selected_Pathways_Weka}{selected pathway using WEKA algorithm generated from lilikoi.select_pathways function}

\item{metaboliteMeasurements}{the metaboliteMeasurements for the samples}

\item{clinical_factors_data}{Metadata}

\item{factors}{which the users want to add to the model}
}
\description{
This function can adjust the best performing model using additional clinical factors specified by the user
It plots ROC for three models: a model building using only the selected pathways, a model built using the clinical factors
and a model built using both the selected pathways and the clinical factors.
}
