% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.machine_learning.r
\name{lilikoi.machine_learning}
\alias{lilikoi.machine_learning}
\title{A machine learning Function}
\usage{
lilikoi.machine_learning(PDSmatrix, measurementLabels, significantPathways)
}
\arguments{
\item{PDSmatrix}{from lilikoi.get_pd_scores}

\item{measurementLabels}{Cancer or not}

\item{significantPathways}{Pathways with high significance from lilikoi.select_pathways function}
}
\description{
This function for classification using 7 different machine learning algorithms
and it plots the ROC curves and the AUC, SEN, and specificity
}
