% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.metab_to_pathway.r
\name{lilikoi.metab_to_pathway}
\alias{lilikoi.metab_to_pathway}
\title{A lilikoi.metab_to_pathway Function}
\usage{
lilikoi.metab_to_pathway(metaboliteNames, searchType)
}
\arguments{
\item{metaboliteNames}{Your metabolite data}

\item{searchType}{The type of the metabolites id such as 'name', 'kegg', 'hmdb','pubchem'}
}
\description{
This function allows you to convert your metabolites id such as names, kegg ids, pubchem ids.
into pathways. Metabolites which have no pathways will be excluded from any downstream analysis.
}
\examples{
matches = lilikoi.metab_to_pathway(c("Asparagine", "Hypotaurine", "5-Oxoproline"), "name")

matches = lilikoi.metab_to_pathway(c("C00152", "C00519", "C01879"), "kegg")

}
