% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilikoi.select_pathways.r
\name{lilikoi.select_pathways}
\alias{lilikoi.select_pathways}
\title{This function will return the significant pathways that can be used for training machine learning models.}
\usage{
lilikoi.select_pathways(PDSmatrix, metaboliteMeasurements, threshold = 0.5,
  method = "info")
}
\arguments{
\item{PDSmatrix}{from lilikoi.get_pd_scores function}

\item{metaboliteMeasurements}{Levels}

\item{threshold}{Cutoff for significance}

\item{method}{Which gain method to use}
}
\description{
This function can reduce the pathway dimension using method specified.
}
