% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loaddata.r
\name{Loaddata}
\alias{Loaddata}
\title{A Loaddata Function}
\usage{
Loaddata(filename)
}
\arguments{
\item{Your}{file name.}
}
\description{
This function allows you to load your metabolomics data.
}
\examples{
Loaddata("data_format.csv")
Loaddata()
}
\keyword{Load}
