% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.r
\name{machine_learning}
\alias{machine_learning}
\title{A machine learning Function}
\usage{
machine_learning(PDSmatrix, selected_Pathways_Weka)
}
\arguments{
\item{PDSmatrix}{from PDSfun function and selected_Pathways_Weka from featuresSelection function}
}
\description{
This function for classification using 7 different machine learning algorithms
and it plot the ROC curves and the AUC, SEN, and specificty
}
\examples{
machine_learning(PDSmatrix,selected_Pathways_Weka)
machine_learning(PDSmatrix,selected_Pathways_Weka)






}
\keyword{classifcation}
