% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adjustment.r
\name{model_adjustment}
\alias{model_adjustment}
\title{Model Adjustemnt function using clinical factors}
\usage{
model_adjustment(result, PDSmatrix, selected_Pathways_Weka,
  clinical_factors_data, factors)
}
\arguments{
\item{PDSmatrix}{the PDS matrix geneerated using PDSfun function}

\item{selected_Pathways_Weka}{selected pathway using WEKA algorithm geneerated from featuresSelection function}

\item{clinical_factors_data}{the metadata for the samples}

\item{factors}{which the users want to add to the model}
}
\description{
This function for adjusted the best performed model using the clinical factors inserted by the user
It plots ROC for three models: model1 build using only seelcted pathways, model2 build using clinical factors
and model3 build using selected pathways and the clinical factors
}
\examples{
model_adjustment(result,PDSmatrix,selected_Pathways_Weka,clinical_factors_data,factors=c('Age','Race'))
model_adjustment(result,PDSmatrix,selected_Pathways_Weka,clinical_factors_data,factors=c('Age','Race'))
}
\keyword{adjustment}
