% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{NameMappingExact}
\alias{NameMappingExact}
\title{A NameMappingExact Function}
\usage{
NameMappingExact(metaboliteDataSet)
}
\arguments{
\item{metaboliteDataSet}{Dataset}
}
\description{
This function matches your metabolites names with databases
with more than 20k metabolites. it allows you also to search among metabolite synonyms.
Make sure that you have the below two databases for the exact matching:
cmpd_db.rda, and syn_nms_db.rda. These should be loaded from lilikoi/R/sysdata.rda
This function was modified version of the name.match function in the below link:
https://github.com/cangfengzhe/Metabo/blob/master/MetaboAnalyst/website/name_match.R
}
\examples{
NameMappingExact()
NameMappingExact()


}
\keyword{Exact}
\keyword{Matching}
