% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{PathMapping}
\alias{PathMapping}
\title{A PathMapping Function}
\usage{
PathMapping(qvec)
}
\arguments{
\item{the}{list of user metabolites and its corrsponding keeg, pubchem and hmdb ids}
}
\description{
This function allows you to map user metabolites to the its corrsponding pathways
make sure that you have the below database for the exact matching:
Sijia_pathway.rda
}
\examples{
PathMapping(qvec)
PathMapping(qvec)
}
\keyword{mapping}
\keyword{pathway}
