% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{PerformApproxMatch}
\alias{PerformApproxMatch}
\title{A PerformApproxMatch Function}
\usage{
PerformApproxMatch(inx)
}
\arguments{
\item{the}{index of unmatched metabolites}
}
\description{
This function allows you to fuzzy matching of your metabolites names with data base
consists of more than 20k metabolites. it allows you also to match also among metabloites synanoms.
make sure that you have the below two databases for the exact matching:
cmpd_db.rda, and syn_nms_db.rda
This function was modified version of the name.match function in the below link:
https://github.com/cangfengzhe/Metabo/blob/master/MetaboAnalyst/website/name_match.R
}
\examples{
PerformApproxMatch(inx)
PerformApproxMatch(inx)


}
\keyword{Matching}
\keyword{fuzzy}
\keyword{logic}
