% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{sanity_check}
\alias{sanity_check}
\title{do some sanity check
A sanity_check Function}
\usage{
sanity_check(name.map, hit.inx)
}
\arguments{
\item{name.map}{is the dataframe of metabolites and its corresponding ids.}

\item{hit.inx}{is the status of the metabolite hits (number if the metabolites matched the database or NA otherwise.)}
}
\description{
This function return how many metabolites matched and unmatched from Metatopathway function
}
\examples{
sanity_check(name.map,hit.inx)
sanity_check


}
\keyword{santiy}
\keyword{search}
